Make import_resources.py more robust.
1. added --clobber argument for overwriting existing files
Defaults to false.

2. use os.path.join for more intelligent path construction.

3. count only resource files that would be (re)generated

+++ _scripts/import_resources.py
@@ -11,7 +11,10 @@ import requests
 RESOURCES_URL="https://docs.google.com/spreadsheets/d/e/2PACX-1vQMEdZXKgYNybkqNv4X26CVNoQZHuE0zb27wuBDgdDwtiyWCICQLhyU_LuLJVeHD5oTRnp-bEdFTuqi/pub?gid=650653420&single=true&output=csv"
 
 parser = argparse.ArgumentParser(description='Parse arguments for this script.')
-parser.add_argument('--dryrun', '-n', default=False, action="store_true")
+parser.add_argument('--dryrun', '-n', default=False, action="store_true",
+                    help='Only print out what would be done.')
+parser.add_argument('--clobber', default=False, action="store_true",
+                    help='Overwrite existing resource files.')
 args = parser.parse_args()
 """Commandline argument values are stored here."""
 
@@ -88,18 +91,22 @@ def main():
         if not has_required:
             continue
 
+        filename = os.path.join(resource_dir, make_row_filename(row))
+        if os.path.isfile(filename) and not args.clobber:
+            print(f"Resource file {filename} already exists, skipping.")
+            continue
+
         rows_accepted += 1
-        filename = f'{resource_dir}{make_row_filename(row)}'
-        
         if args.dryrun:
             print(f"*** Dry run *** {filename} would contain:")
             print(format_row_contents(row))
-        else:
-            with open(filename, encoding='utf-8', mode='w+') as out_file:
-                out_file.write(format_row_contents(row))
-                out_file.close()
+            continue
 
-    print(f'Imported {rows_accepted} resources.')
+        with open(filename, encoding='utf-8', mode='w+') as out_file:
+            out_file.write(format_row_contents(row))
+            out_file.close()
+
+    print(f'(Re)generated {rows_accepted} resource files.')
         
 
 if __name__ == '__main__':
